% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/and.R
\name{and}
\alias{and}
\alias{or}
\title{Combine a vector into a natural language string}
\usage{
and(x, ..., lang = NULL)

or(x, ..., lang = NULL)
}
\arguments{
\item{x}{A list of \link{character} strings to combine}

\item{...}{These dots are for future extensions and must be empty.}

\item{lang}{The language to use. If \code{\link{NULL}}, the default, the language of
the user's R session is used.}
}
\value{
A \link{character} string of length 1
}
\description{
These functions transform a vector into a single string similar to
\code{\link[knitr:combine_words]{knitr::combine_words()}} or \code{\link[glue:glue_collapse]{glue::glue_collapse()}}. \code{and()} and \code{or()}
natively support localization, using translations and punctuation to match
the users' language settings. (See section "Language" for more details.)
\itemize{
\item \code{and()} combines words using the native conjunctive ("and" in English)
\item \code{or()} combines words using the native disjunctive ("or" in English)
}
}
\section{Languages}{


Translations are available in the following languages:\tabular{llllll}{
   \strong{code} \tab \strong{language} \tab  \tab \strong{and} \tab \strong{or} \tab  \cr
   ca \tab català \tab Catalan \tab i \tab o \tab  \cr
   de \tab Deutsch \tab German \tab und \tab oder \tab  \cr
   en \tab English \tab  \tab and \tab or \tab  \cr
   en_US \tab US English \tab  \tab and \tab or \tab (uses Oxford comma) \cr
   eo \tab esperanto \tab Esperanto \tab kaj \tab aŭ \tab  \cr
   es \tab español \tab Spanish \tab y / e \tab o / u \tab  \cr
   eu \tab euskara \tab Basque \tab eta \tab edo \tab  \cr
   fr \tab français \tab French \tab et \tab ou \tab  \cr
   it \tab italiano \tab Italian \tab e / ed \tab o / od \tab  \cr
   pt \tab português \tab Portuguese \tab e \tab ou \tab  \cr
   tr \tab Türkçe \tab Turkish \tab ve \tab veya \tab  \cr
}
}

\examples{
and(c("Betty", "James", "Inez"))
or(c("win", "lose", "draw"))

and(c("libre", "soberana", "independiente"), lang = "es")
and(c("liberté", "égalité", "fraternité"), lang = "fr")

}
