% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/and_more.R
\name{and_more}
\alias{and_more}
\title{Combine a vector into a natural language string with a limited length}
\usage{
and_more(x, max = 3, ..., lang = NULL)
}
\arguments{
\item{x}{A list of \link{character} strings to combine}

\item{max}{}

\item{...}{These dots are for future extensions and must be empty.}

\item{lang}{The language to use. If \code{\link{NULL}}, the default, the language of
the user's R session is used.}
}
\value{
A \link{character} string of length 1
}
\description{
These functions combine a specified number of elements of a vector into a
single string indicating how many more elements are in the vector.
\code{and_more()} and \code{or_more()} natively support localization, using
translations and punctuation to match the users' language settings.
(See section "Language" for more details.)
\itemize{
\item \code{and_more()} combines words using the native conjunctive ("and" in English)
\item \code{or_more()} combines words using the native disjunctive ("or" in English)
}
}
\section{Languages}{


Translations are available in the following languages:\tabular{llll}{
   \strong{code} \tab \strong{language} \tab  \tab \strong{and ... more} \cr
   ca \tab català \tab Catalan \tab i ... més \cr
   de \tab Deutsch \tab German \tab und ... weitere \cr
   en \tab English \tab  \tab and ... more \cr
   en_US \tab US English \tab  \tab and ... more \cr
   eo \tab esperanto \tab Esperanto \tab kaj ... pli \cr
   es \tab español \tab Spanish \tab y ... más \cr
   eu \tab euskara \tab Basque \tab eta ... gehiago \cr
   fr \tab français \tab French \tab et 1 de plus / et ... autres \cr
   it \tab italiano \tab Italian \tab e 1 altro / e ... altri \cr
   pt \tab português \tab Portuguese \tab e mais ... \cr
   tr \tab Türkçe \tab Turkish \tab ve ... tane daha \cr
}
}

\examples{
and_more(letters)
or_more(letters)

and_more(letters, max = 10)

and_more(letters, lang = "es")
and_more(letters, lang = "fr")

}
