% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/and-package.R
\docType{package}
\name{and-package}
\alias{and-package}
\alias{_PACKAGE}
\alias{reprex-package}
\title{and: Construct Natural-Language Lists with Internationalization}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Construct language-aware lists. Make "and"-separated and "or"-separated lists that automatically conform to the user's language settings.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://and.rossellhayes.com}
  \item \url{https://github.com/rossellhayes/and}
  \item Report bugs at \url{https://github.com/rossellhayes/and/issues}
}

}
\author{
\strong{Maintainer}: Alexander Rossell Hayes \email{alexander@rossellhayes.com} (\href{https://orcid.org/0000-0001-9412-0457}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item  Unicode, Inc. (Language data) [data contributor]
}

}
\keyword{internal}
