% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/and_languages.R
\docType{data}
\name{and_languages}
\alias{and_languages}
\title{Supported languages}
\format{
A data frame with 6 variables:
\describe{
\item{language}{The name of the language, possibly with a territory in parentheses}
\item{code}{The language code}
\item{example_and_2}{An example of a conjunctive list with two elements in the language}
\item{example_and_4}{An example of a conjunctive list with four elements in the language}
\item{example_or_2}{An example of a disjunctive list with two elements in the language}
\item{example_or_4}{An example of a disjunctive list with four elements in the language}
\item{support}{Either "full" or "partial". Partially supported languages generally localize \code{\link[=and]{and()}} but not \code{\link[=or]{or()}}.}
}
}
\source{
Language data is derived from the
\href{https://cldr.unicode.org/}{Unicode Common Locale Data Repository (CLDR)}
}
\usage{
and_languages
}
\description{
A list of supported languages and examples of their usage.
}
\keyword{datasets}
